requirements.txt
pyautogen>=0.2.0
termcolor>=2.3.0
prompt_toolkit>=3.0.43 